NFC: Make a few loop transformations more accessible
This moves `FuseForallOp` to the IREEGPU dialect because it generates
IREEGPU operations, avoiding a global dependency on the dialect in
codegen. The `HoistForallFromFor` pattern is moved to
`Codegen/Transforms` because it applies only to operations used across
all codegen backends.
