+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CommonExtensions.cpp
@@ -182,220 +182,9 @@ void transform_dialect::ApplyFoldFillIntoPadPatternsOp::populatePatterns(
 // ApplyHoistForallFromForPatternsOp
 //===---------------------------------------------------------------------===//
 
-namespace {
-struct HoistForallFromFor : public OpRewritePattern<scf::ForOp> {
-  using OpRewritePattern<scf::ForOp>::OpRewritePattern;
-  LogicalResult matchAndRewrite(scf::ForOp loop,
-                                PatternRewriter &rewriter) const final {
-    if (loop.getBody()->getOperations().size() != 2) {
-      return rewriter.notifyMatchFailure(
-          loop, "Body of the loop contains more than one op");
-    }
-
-    // TODO(qedawkins): It should be fine to hoist as long as there is a single
-    // forall op such that any operation within the block of the parent scf.for
-    // is independent of the destination of the forall.
-    auto forallOp =
-        dyn_cast<scf::ForallOp>(loop.getBody()->without_terminator().begin());
-    if (!forallOp) {
-      return rewriter.notifyMatchFailure(
-          loop, "Loop single contained op is not scf.forall op");
-    }
-
-    if (forallOp.getNumResults() != 1 || loop.getNumResults() != 1) {
-      return rewriter.notifyMatchFailure(
-          loop, "unimplemented: multi-result hoisting");
-    }
-
-    if (loop.getBody()->getTerminator()->getOperand(0).getDefiningOp() !=
-        forallOp) {
-      return rewriter.notifyMatchFailure(
-          loop, "Single for loop return is not forall result");
-    }
-
-    // Step 1. Collect the set of tensor.parallel_insert_slice ops in the
-    // terminator and their paired extract_slice ops from the for loop iter arg.
-    SmallVector<Operation *> sliceOperandProducers;
-
-    BackwardSliceOptions backwardOptions;
-    backwardOptions.inclusive = true;
-    backwardOptions.filter = [&](Operation *op) -> bool {
-      return forallOp->isProperAncestor(op);
-    };
-    SetVector<Operation *> slice;
-
-    scf::InParallelOp parallelTerminator = forallOp.getTerminator();
-    SmallVector<tensor::ParallelInsertSliceOp> terminators(
-        forallOp.getNumResults());
-    SmallVector<tensor::ExtractSliceOp> pairedSlices(forallOp.getNumResults());
-    int64_t numInductionVars = forallOp.getInductionVars().size();
-    for (auto &yieldingOp : parallelTerminator.getYieldingOps()) {
-      auto parallelInsert = cast<tensor::ParallelInsertSliceOp>(&yieldingOp);
-      BlockArgument destBbArg =
-          llvm::cast<BlockArgument>(parallelInsert.getDest());
-      tensor::ExtractSliceOp destSlice;
-      for (auto user : destBbArg.getUsers()) {
-        if (user == parallelInsert)
-          continue;
-        auto maybeSlice = dyn_cast<tensor::ExtractSliceOp>(user);
-        // Fail if the destination has more users than a direct insert and
-        // extract slice.
-        if (!maybeSlice) {
-          return failure();
-        }
-        // Require a single extract per destination.
-        if (destSlice) {
-          return failure();
-        }
-        destSlice = maybeSlice;
-      }
-      // Verify they operate on equivalent subsets, ensuring the slices are
-      // hoistable. It is still possible to hoist the loop if this is not true,
-      // however in such cases we likely formed the loops in the wrong order.
-      if (!cast<SubsetOpInterface>(*destSlice)
-               .operatesOnEquivalentSubset(
-                   cast<SubsetOpInterface>(*parallelInsert),
-                   [](Value v1, Value v2) { return v1 == v2; })) {
-        return failure();
-      }
-      terminators[destBbArg.getArgNumber() - numInductionVars] = parallelInsert;
-      pairedSlices[destBbArg.getArgNumber() - numInductionVars] = destSlice;
-
-      // Collect all of the offset/size/stride operands for both slices and
-      // compute a backwards slice of the program from them. Fail if any of
-      // them depend on the serial loop iterator.
-      llvm::SmallDenseSet<Value> sliceOperands;
-      sliceOperands.insert(
-          parallelInsert.getOperands().begin() +
-              parallelInsert.getOffsetSizeAndStrideStartOperandIndex(),
-          parallelInsert.getOperands().end());
-      sliceOperands.insert(
-          destSlice.getOperands().begin() +
-              destSlice.getOffsetSizeAndStrideStartOperandIndex(),
-          destSlice.getOperands().end());
-      for (Value operand : sliceOperands) {
-        if (auto bbArg = dyn_cast<BlockArgument>(operand)) {
-          if (bbArg.getOwner()->getParentOp() == loop) {
-            return rewriter.notifyMatchFailure(
-                loop, "Slice operand producers depend on loop");
-          }
-        }
-        SetVector<Operation *> tmpBackwardSlice;
-        getBackwardSlice(operand, &tmpBackwardSlice, backwardOptions);
-        slice.set_union(tmpBackwardSlice);
-      }
-    }
-
-    // An operation is safe to hoist if it is speculatable and none of its
-    // operands depend on the serial loop.
-    auto isSafeToHoist = [&](Operation *op) {
-      return op->getBlock() == forallOp.getBody() && isMemoryEffectFree(op) &&
-             isSpeculatable(op) &&
-             llvm::none_of(op->getOperands(), [&](Value operand) {
-               if (auto blockArg = dyn_cast<BlockArgument>(operand)) {
-                 return blockArg.getOwner() == loop.getBody();
-               }
-               return false;
-             });
-    };
-
-    if (!llvm::all_of(slice, isSafeToHoist)) {
-      return rewriter.notifyMatchFailure(
-          loop, "Slice operand producers not safe to hoist out of loop");
-    }
-
-    // Sort the backwards slice of the producers for the insertion/extraction
-    // indices by the block order in the scf.forall body. This ensures that we
-    // hoist operations in the same order they started. Any topological ordering
-    // would work too because the operations are speculatable.
-    slice = mlir::topologicalSort(slice);
-
-    // Step 2. Create the ForallOp.
-    Location loc = forallOp.getLoc();
-    scf::ForallOp newForallOp = rewriter.create<scf::ForallOp>(
-        loc, forallOp.getMixedLowerBound(), forallOp.getMixedUpperBound(),
-        forallOp.getMixedStep(), loop.getInitArgs(), forallOp.getMappingAttr());
-
-    {
-      // RAII guard, inserting within forallOp, before terminator.
-      OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPoint(newForallOp.getTerminator());
-      SmallVector<Value> newInits;
-      for (auto slice : pairedSlices) {
-        newInits.push_back(slice.getResult());
-      }
-      // Step 3. Create a new for loop with new inits for the result of the
-      // extracted slices.
-      auto newLoop = rewriter.create<scf::ForOp>(
-          loop.getLoc(), loop.getLowerBound(), loop.getUpperBound(),
-          loop.getStep(), newInits,
-          [](OpBuilder &, Location, Value, ValueRange) {});
-
-      {
-        // Step 4. Inline the body of the original forall into the new for loop.
-        OpBuilder::InsertionGuard g2(rewriter);
-        SmallVector<Value> argReplacements(newForallOp.getInductionVars());
-        argReplacements.append(newForallOp.getRegionIterArgs().begin(),
-                               newForallOp.getRegionIterArgs().end());
-        rewriter.mergeBlocks(forallOp.getBody(), newLoop.getBody(),
-                             argReplacements);
-        rewriter.replaceAllUsesWith(loop.getInductionVar(),
-                                    newLoop.getInductionVar());
-        // Replace the users of the to-be-hoisted slices with the loop iter
-        // args.
-        for (auto [hoistedSlice, iterArg] :
-             llvm::zip_equal(pairedSlices, newLoop.getRegionIterArgs())) {
-          rewriter.replaceAllUsesExcept(hoistedSlice, iterArg, newLoop);
-        }
-
-        // Create the terminator for the new loop using the sources of the
-        // parallel inserts.
-        SmallVector<Value> newYields;
-        for (auto parallelSlice : terminators) {
-          newYields.push_back(parallelSlice.getSource());
-        }
-        rewriter.setInsertionPointToEnd(newLoop.getBody());
-        rewriter.create<scf::YieldOp>(loop.getLoc(), newYields);
-      }
-
-      // Move all producers for the indices of the slices outside of the body
-      // of the loop (and the extract_slice ops themselves).
-      for (auto sliceOperandProducer : slice) {
-        rewriter.moveOpBefore(sliceOperandProducer, newLoop);
-      }
-      for (auto slice : pairedSlices) {
-        rewriter.moveOpBefore(slice, newLoop);
-      }
-
-      // Create the new terminator for the hoisted forall loop using the results
-      // of the new for loop.
-      rewriter.setInsertionPointToEnd(newForallOp.getTerminator().getBody());
-      for (auto [parallelSlice, source, dest] :
-           llvm::zip_equal(terminators, newLoop.getResults(),
-                           newForallOp.getRegionIterArgs())) {
-        rewriter.create<tensor::ParallelInsertSliceOp>(
-            parallelSlice.getLoc(), source, dest, parallelSlice.getOffsets(),
-            parallelSlice.getSizes(), parallelSlice.getStrides(),
-            parallelSlice.getStaticOffsets(), parallelSlice.getStaticSizes(),
-            parallelSlice.getStaticStrides());
-      }
-    }
-
-    // Step 5. Erase the original terminator and replace the loop with
-    // the hoisted loop.
-    for (auto parallelSlice : terminators) {
-      rewriter.eraseOp(parallelSlice);
-    }
-    rewriter.eraseOp(parallelTerminator);
-    rewriter.replaceOp(loop, newForallOp);
-    return success();
-  }
-};
-} // namespace
-
 void transform_dialect::ApplyHoistForallFromForPatternsOp::populatePatterns(
     RewritePatternSet &patterns) {
-  patterns.insert<HoistForallFromFor>(patterns.getContext());
+  iree_compiler::populateForallLoopHoistingPattern(patterns);
 }
 
 //===---------------------------------------------------------------------===//
@@ -896,201 +685,6 @@ DiagnosedSilenceableFailure transform_dialect::ForallToWorkgroupOp::applyToOne(
 void transform_dialect::ForallToWorkgroupOp::getEffects(
     SmallVectorImpl<MemoryEffects::EffectInstance> &effects) {
   transform::onlyReadsHandle(getTarget(), effects);
-  transform::modifiesPayload(effects);
-}
-
-//===---------------------------------------------------------------------===//
-// FuseForallOp
-//===---------------------------------------------------------------------===//
-
-static FailureOr<int64_t> getTripCount(scf::ForallOp loop) {
-  ArrayRef<int64_t> lbs = loop.getStaticLowerBound();
-  ArrayRef<int64_t> ubs = loop.getStaticUpperBound();
-  ArrayRef<int64_t> steps = loop.getStaticStep();
-
-  if (ShapedType::isDynamicShape(lbs) || ShapedType::isDynamicShape(ubs) ||
-      ShapedType::isDynamicShape(steps)) {
-    return failure();
-  }
-
-  int64_t tripCount = 1;
-  for (auto [lb, ub, step] : llvm::zip_equal(lbs, ubs, steps)) {
-    tripCount *= mlir::ceilDiv((ub - lb), step);
-  }
-  return tripCount;
-}
-
-static LogicalResult compareWorkerCountsAndTypes(scf::ForallOp producer,
-                                                 scf::ForallOp consumer) {
-  FailureOr<int64_t> producerTripCount = getTripCount(producer);
-  FailureOr<int64_t> consumerTripCount = getTripCount(consumer);
-  if (failed(producerTripCount) || failed(consumerTripCount) ||
-      *producerTripCount != *consumerTripCount) {
-    return failure();
-  }
-
-  auto checkMappingTypes = [&](ArrayAttr array) {
-    return llvm::all_of(array.getValue(),
-                        llvm::IsaPred<gpu::GPUThreadMappingAttr>) ||
-           llvm::all_of(array.getValue(),
-                        llvm::IsaPred<gpu::GPUWarpMappingAttr>);
-  };
-
-  if (producer.getMappingAttr() != consumer.getMappingAttr() ||
-      !checkMappingTypes(producer.getMappingAttr()) ||
-      !checkMappingTypes(consumer.getMappingAttr())) {
-    return failure();
-  }
-  return success();
-}
-
-static void replaceExtractSlice(RewriterBase &rewriter, Location loc,
-                                tensor::ParallelInsertSliceOp parallelInsert,
-                                tensor::ExtractSliceOp extractSlice) {
-  OpBuilder::InsertionGuard g(rewriter);
-  auto shuffleOp = rewriter.create<IREE::GPU::ShuffleTensorOp>(
-      loc, extractSlice.getType(), parallelInsert.getSource(),
-      parallelInsert.getOffsets(), parallelInsert.getSizes(),
-      parallelInsert.getStrides(), parallelInsert.getStaticOffsets(),
-      parallelInsert.getStaticSizes(), parallelInsert.getStaticStrides(),
-      parallelInsert.getDest());
-  Region *region = &shuffleOp.getRegion();
-  rewriter.createBlock(region, region->end(),
-                       ArrayRef<Type>{parallelInsert.getDestType()},
-                       ArrayRef<Location>{loc});
-  rewriter.setInsertionPointToStart(shuffleOp.getBody());
-  auto terminator =
-      rewriter.create<IREE::GPU::YieldOp>(loc, extractSlice.getResult());
-  rewriter.moveOpBefore(extractSlice, terminator);
-  extractSlice.getSourceMutable().assign(shuffleOp.getBody()->getArgument(0));
-  rewriter.replaceAllUsesExcept(extractSlice.getResult(), shuffleOp,
-                                terminator);
-}
-
-LogicalResult fuseForallIntoSlice(RewriterBase &rewriter,
-                                  scf::ForallOp producer,
-                                  scf::ForallOp consumer,
-                                  tensor::ExtractSliceOp slice,
-                                  std::optional<Attribute> addressSpace) {
-  if (producer->getNumResults() != 1) {
-    return failure();
-  }
-
-  if (failed(compareWorkerCountsAndTypes(producer, consumer))) {
-    return failure();
-  }
-
-  auto isAll = [](ArrayRef<OpFoldResult> array, int64_t cmp) {
-    return llvm::all_of(array, [cmp](OpFoldResult val) {
-      return isConstantIntValue(val, cmp);
-    });
-  };
-
-  if (!isAll(producer.getMixedStep(), 1) ||
-      !isAll(producer.getMixedLowerBound(), 0) ||
-      !isAll(consumer.getMixedStep(), 1) ||
-      !isAll(consumer.getMixedLowerBound(), 0)) {
-    return failure();
-  }
-
-  rewriter.setInsertionPoint(slice);
-
-  // Step 1. Compute the producer IDs in terms of the consumer IDs.
-
-  MLIRContext *context = rewriter.getContext();
-  Location loc = producer.getLoc();
-
-  AffineExpr d0, d1, d2;
-  bindDims(context, d0, d1, d2);
-  AffineExpr mulAdd = d0 * d1 + d2;
-  OpFoldResult linearId = rewriter.getIndexAttr(0);
-  for (auto [inductionVar, workerCount] :
-       llvm::zip_equal(getAsOpFoldResult(consumer.getInductionVars()),
-                       consumer.getMixedUpperBound())) {
-    linearId = affine::makeComposedFoldedAffineApply(
-        rewriter, loc, mulAdd, {linearId, workerCount, inductionVar});
-  }
-
-  Value linearThreadIdVal =
-      getValueOrCreateConstantIndexOp(rewriter, loc, linearId);
-  SmallVector<Value> ranges;
-  for (auto workerCount : producer.getStaticUpperBound()) {
-    ranges.push_back(rewriter.create<arith::ConstantIndexOp>(loc, workerCount));
-  }
-  ValueRange newIds = rewriter
-                          .create<affine::AffineDelinearizeIndexOp>(
-                              loc, linearThreadIdVal, ranges)
-                          .getResults();
-
-  // Step 2. Inline the region of the producer.
-  SmallVector<Value> bbArgReplacements(newIds);
-  bbArgReplacements.append(producer.getOutputs().begin(),
-                           producer.getOutputs().end());
-
-  scf::InParallelOp terminator = producer.getTerminator();
-  rewriter.inlineBlockBefore(producer.getBody(), slice, bbArgReplacements);
-
-  rewriter.setInsertionPointAfter(terminator);
-  auto parallelInsert =
-      cast<tensor::ParallelInsertSliceOp>(*terminator.getYieldingOps().begin());
-
-  replaceExtractSlice(rewriter, loc, parallelInsert, slice);
-
-  rewriter.eraseOp(parallelInsert);
-  rewriter.eraseOp(terminator);
-  rewriter.eraseOp(producer);
-  return success();
-}
-
-DiagnosedSilenceableFailure
-transform_dialect::FuseForallOp::apply(transform::TransformRewriter &rewriter,
-                                       transform::TransformResults &results,
-                                       transform::TransformState &state) {
-  auto producers = state.getPayloadOps(getProducer());
-  auto consumers = state.getPayloadOps(getConsumer());
-
-  int64_t numProducers = llvm::range_size(producers);
-  int64_t numConsumers = llvm::range_size(consumers);
-  if (numProducers != 1 || numConsumers != 1) {
-    return mlir::emitDefiniteFailure(state.getTopLevel(),
-                                     "More than one producer or consumer");
-  }
-
-  auto producer = dyn_cast<scf::ForallOp>(*producers.begin());
-  auto consumer = dyn_cast<scf::ForallOp>(*consumers.begin());
-  if (!producer || !consumer) {
-    return mlir::emitDefiniteFailure(state.getTopLevel(),
-                                     "Non-forall producer or consumer");
-  }
-
-  if (!producer->hasOneUse()) {
-    return mlir::emitDefiniteFailure(state.getTopLevel(),
-                                     "non-single use producer");
-  }
-
-  auto sliceConsumer =
-      dyn_cast<tensor::ExtractSliceOp>(*producer->user_begin());
-  if (!sliceConsumer || sliceConsumer->getParentOp() != consumer) {
-    return mlir::emitDefiniteFailure(state.getTopLevel(),
-                                     "producer loop sole consumer is not an "
-                                     "extracted slice from the consumer loop");
-  }
-
-  if (failed(fuseForallIntoSlice(rewriter, producer, consumer, sliceConsumer,
-                                 getAddressSpace()))) {
-    return mlir::emitDefiniteFailure(state.getTopLevel(),
-                                     "failed to fuse forall ops");
-  }
-
-  results.set(getOperation()->getOpResult(0), {consumer});
-  return DiagnosedSilenceableFailure::success();
-}
-
-void transform_dialect::FuseForallOp::getEffects(
-    SmallVectorImpl<MemoryEffects::EffectInstance> &effects) {
-  transform::consumesHandle(getProducer(), effects);
-  transform::consumesHandle(getConsumer(), effects);
-  transform::producesHandle(getResult(), effects);
   transform::modifiesPayload(effects);
 }
