+++ compiler/src/iree/compiler/Codegen/Common/test/BUILD.bazel
@@ -67,7 +67,6 @@ iree_lit_test_suite(
             "transform_copy_operand.mlir",
             "transform_distribute_lane_forall.mlir",
             "transform_flatten_forall.mlir",
-            "transform_fuse_forall.mlir",
             "transform_hoist_forall.mlir",
             "transform_lower_shuffle.mlir",
             "transform_match_partial_reduction.mlir",
