+++ compiler/src/iree/compiler/Codegen/Common/test/CMakeLists.txt
@@ -63,7 +63,6 @@ iree_lit_test_suite(
     "transform_copy_operand.mlir"
     "transform_distribute_lane_forall.mlir"
     "transform_flatten_forall.mlir"
-    "transform_fuse_forall.mlir"
     "transform_hoist_forall.mlir"
     "transform_lower_shuffle.mlir"
     "transform_match_partial_reduction.mlir"
