+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/Transforms/Transforms.h
@@ -14,6 +14,8 @@
 #define IREE_COMPILER_CODEGEN_DIALECT_GPU_TRANSFORMS_TRANSFORMS_H_
 
 #include "iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUOps.h"
+#include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
 #include "mlir/Dialect/Vector/Transforms/VectorRewritePatterns.h"
 #include "mlir/IR/PatternMatch.h"
 
@@ -23,8 +25,18 @@ struct UnrollVectorOptions;
 
 namespace mlir::iree_compiler::IREE::GPU {
 
+/// Function to fuse the given producer-consumer pair of forall loops into
+/// the single consumer loop at the given |slice| within the consumer of the
+/// producer. This is managed by inserting an `iree_gpu.tensor_shuffle` at the
+/// boundary to synchronize the workers at the fusion point.
+LogicalResult fuseForallIntoSlice(RewriterBase &rewriter,
+                                  scf::ForallOp producer,
+                                  scf::ForallOp consumer,
+                                  tensor::ExtractSliceOp slice);
+
 void populateIREEGPUVectorUnrollPatterns(
     RewritePatternSet &patterns, const vector::UnrollVectorOptions &options);
+
 void populateIREEGPUVectorizationPatterns(RewritePatternSet &patterns);
 
 } // namespace mlir::iree_compiler::IREE::GPU
